Site Pages, Admin Pages & Trees, System Settings
This large update includes tons of upgrades related to running multiple
form-generating-branching-trees in one installation. Pages (which are a
type of Tree) can now be setup for the public visitors and
administrators, and a WYSIWYG has been integrated (for now, Summernote,
but as always, I’m open to suggestions). The SurvLoop-related sections
of the admin menu have also been slimmed down, and the System Settings
can finally be edited through a basic needs page on the admin side.